refactor search functions
moves search functions to its own file and provides a struct common for search/replace operations. the struct is embedded with commands that utilize search so that the commands contain the search paramters and the commands implement an interface for retrieving the parameters. this allows common methods to be located on a `Searcher` struct that can reference the command via the interface.