refactor search and implement replace command (#69)
* Add initial test for replace command

* Add replace command scaffold

* refactor search functions

moves search functions to its own file and provides a struct common for search/replace operations. the struct is embedded with commands that utilize search so that the commands contain the search paramters and the commands implement an interface for retrieving the parameters. this allows common methods to be located on a `Searcher` struct that can reference the command via the interface.

* implement replace command

the `replace` command will do a match in the same fashion as `grep` but instead of presenting results with search patterns highlighted, it will highlight matches in red followed immediately by replacement string in green. in this way, it is easy to identify on one line what will be matched and replaced. by default, the operation will ask for confirmation before writing back to vault. it is possible to use flags `-y` to write without confirmation, `-n` to skip confirmation and not write aka dry-run, and limit scope to keys (`-k`) and values (`-v`) in the same way as in `grep`.

* add a few more replace assertions

just want to make sure that the replace operation is not wiping out the rest of the secret.

Co-authored-by: Karl Fischer <fishi0x01@gmail.com>