add shallow flag to grep/replace commands (#97)
this flag limits searches to only the immediate leaf nodes of a path rather than deeply recursing through every path below the specified path. we found this to be very useful when we have a large path structure below a parent path and want to exclude it during these operations.